backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
feedparser==6.0.11
requests==2.32.3
APScheduler==3.11.0
